feat: change default behaviour
this PR aims to change the default behaviour of the package:

In case the package gets called the first time ever in a dapp **and** only one wallet is installed **instead** of showing a list with one wallet installed + discovery, **replace by** picking the one installed wallet by default.
This is the default behaviour in @argent/get-starknet currently and doesnt force the user to do one additional step on every initial dapp usage.