feat: change Provider to ProviderInterface in IStarknetWindowObject
That would make more sense if we replace the type of `provider` in IStarknetWindowObject
to ProviderInterface instead of Provider. I assume there is a reason why it has not been
done because the account already relies on the interface! Can you explain? If you have
the need to a specific method which does not look to be the case since the project compile
that would make sense to add it to the interface in starknet.js. I am working on an
implementation of a StarknetWindowObject and relying on the defaultProvider inflates my
artefact. Thank you for your feedback.